LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/width-expected.txt


PASS Can set 'width' to CSS-wide keywords
FAIL Can set 'width' to var() references assert_equals: expected 2 but got 1
PASS Can set 'width' to the 'auto' keyword
FAIL Can set 'width' to a percent assert_approx_equals: expected -3.14 +/- 0.000001 but got 0
FAIL Can set 'width' to a length assert_equals: expected "px" but got "em"
PASS Setting 'width' to a time throws TypeError
PASS Setting 'width' to an angle throws TypeError
PASS Setting 'width' to a flexible length throws TypeError
FAIL Setting 'width' to a number throws TypeError assert_throws_js: function "() => styleMap.set(propertyName, example.input)" did not throw
PASS Setting 'width' to a URL throws TypeError
PASS Setting 'width' to a transform throws TypeError
PASS Can set 'min-width' to CSS-wide keywords
FAIL Can set 'min-width' to var() references assert_equals: expected 2 but got 1
FAIL Can set 'min-width' to a percent assert_approx_equals: expected -3.14 +/- 0.000001 but got 0
FAIL Can set 'min-width' to a length assert_equals: expected "px" but got "em"
PASS Setting 'min-width' to a time throws TypeError
PASS Setting 'min-width' to an angle throws TypeError
PASS Setting 'min-width' to a flexible length throws TypeError
FAIL Setting 'min-width' to a number throws TypeError assert_throws_js: function "() => styleMap.set(propertyName, example.input)" did not throw
PASS Setting 'min-width' to a URL throws TypeError
PASS Setting 'min-width' to a transform throws TypeError
PASS Can set 'max-width' to CSS-wide keywords
FAIL Can set 'max-width' to var() references assert_equals: expected 2 but got 1
PASS Can set 'max-width' to the 'none' keyword
FAIL Can set 'max-width' to a percent Invalid values
FAIL Can set 'max-width' to a length Invalid values
PASS Setting 'max-width' to a time throws TypeError
PASS Setting 'max-width' to an angle throws TypeError
PASS Setting 'max-width' to a flexible length throws TypeError
FAIL Setting 'max-width' to a number throws TypeError assert_throws_js: function "() => styleMap.set(propertyName, example.input)" did not throw
PASS Setting 'max-width' to a URL throws TypeError
PASS Setting 'max-width' to a transform throws TypeError
